requirements.txt
langchain
gradio
openai
python-dotenv